Compilation fixes
Former-commit-id: bfb16585572dd377658b4d96eb2954693886bdc3

+++ hq-server/src/main/java/org/hyperic/hq/inventory/domain/IdentityAware.java
@@ -1,6 +1,6 @@
 package org.hyperic.hq.inventory.domain;
 
 public interface IdentityAware {
-	public Long getId();
-	public void setId(Long id);
+	public Integer getId();
+	public void setId(Integer id);
 }
