+++ hq-web/src/main/java/org/hyperic/hq/api/Representation.java
@@ -49,7 +49,7 @@ public class Representation {
 			this.data = data;
 			
 			if (data instanceof IdentityAware) {
-				Long id = ((IdentityAware) data).getId();
+				Integer id = ((IdentityAware) data).getId();
 				
 				this.links.put("self", instanceUri.expand(domain.toString(), id).toASCIIString());
